Fix broken alignment in src/slab.c
Alignment macro was buggy, thus resulting in bad alignment.
This means speed hit on x86, but may result in crashes
on arches which are strict on alignment, unless the macro
is pre-defined in OS headers.

No problems from the field have been reported, thus maybe
even the bad alignment happened to be aligned?