Don't release server during unnamed extended query statement.
Some drivers do parse/describe/sync followed by bind/execute ... sync.
Unfortunately, the first sync results in a Ready For Query response, and
if we're not in a transaction the server is released. This makes
transaction mode close to useless for these drivers unless every
statement occurs in the context of an explicit transaction. To overcome
this, we detect when a parse is for the unnamed statement, and then
don't mark the server as ready until we have seen either a subsequent
bind or a simple query. To do that we add an extra boolean state to the
server called wait_for_bind.