doc: use rst2man for manpages
Although Asciidoc has nice markup and nice output,
it depends on DocBook & TeX, which makes it rather
annoying to have around.

Website is now based on Markdown, but it's not fit for
manpage conversion - not enough syntax.  It is really
meant for only HTML production, as it expect HTML fallbacks
for anything.

So use rst2man for manpages, it makes docs readable on github
too and via some hacks can be converted to markdown.