Consistent search_path between client and server
This enables consistent search_path caching on the client, which is
applied to the server each time a client is assigned a server socket.

Every time a client sends a 'SET SEARCH_PATH...', the client->vars are
updated with the current search_path value. This value is then set on
a server backend each time that client is reassigned a server.

Note: search_path is not cached on the client if the pool_mode is set
to POOL_SESSION as it would be superfluous to do so, given that there
is a 1:1 mapping between client->server.

This also adds output to 'SHOW FDS' to allow viewing of currently
cached search_path values on client fds.