Handle auth_type=password when the stored password is md5
When used in the typical way (by querying pg_shadow), auth_query or userlist
generation will return md5 hashed passwords. Under such a situation, we should
still allow clients to authenticate with unhashed passwords. We should also
correctly handle the situation where we have an md5 hashed password, but the
server is expecting a plaintext one. We do this by caching passwords that we
have determined match the expected hash. We can then send those passwords
upstream to a server.