Log error on failure to load auth_file
Only logging a debug message when things work and not an error when they
don't is a bit user-unfriendly.

Explicitly avoid logging in the case that auth_file hasn't been
configured at all, to make sure we don't spam logs unnecessraily. But if
a configured file exists and can't be opened, we log.

Fixes #192