Allow setting per-user connection limits to unlimited
See https://github.com/pgbouncer/pgbouncer/issues/166 for rationale.

There is undocumented support for per-user connection limits already in
pgBouncer. However it is not quite flexible enough for us. We wanted to be
able to put a limit on all users, except for specific users (e.g. DBAs, system
processes) which are unlimited. So we hoped to have a low setting for the
global `max_user_connections`, and override it for specific users to
*unlimited* in `pgbouncer.ini:`

	max_user_connections=3

	[users]
	system = max_user_connections=0
	dba = max_user_connections=0

This patch makes it possible to set the per-user connection limit to either `0`
(unlimited) or `-1` (fallback to global default), or a specific limit as before.

Also, PAM users will have their limits copied from non-PAM users of the same
name in the [users] section.