Improve logging when per-user connection limit is hit
See https://github.com/pgbouncer/pgbouncer/issues/166 for rationale.

It's difficult to tell from the log messages when a user is being made to wait
for a free connection in pgBouncer itself.  The logs would normally say:

	C-0x692020: testing2/testuser@1.2.3.4:33034 login attempt: db=testing2 user=testuser tls=no
	S-0x696cd0: testing2/testuser@127.0.0.1:5433 new connection to server (from 127.0.0.1:38388)

Where the second line indicates a new connection to the server. This line would
not appear in one of two cases:

* An existing connection is reused, or
* No connection is available and the user must wait for one to become free.

This patch adds extra logging when a user is made to wait for a connection to
become free. For example:

	C-0x692020: testing2/testuser@1.2.3.4:33034 login attempt: db=testing2 user=testuser tls=no
	S-0x696cd0: testing2/testuser@127.0.0.1:5433 new connection to server (from 127.0.0.1:38388)
	C-0x6921a8: testing2/testuser@1.2.3.4:33038 login attempt: db=testing2 user=testuser tls=no
	C-0x6921a8: testing2/testuser@1.2.3.4:33038 per_loop_activate: reached per-user connection limit, waiting for one to become free
	C-0x691d10: testing2/testuser@1.2.3.4:33026 closing because: client unexpected eof (age=23)
	S-0x6969c0: testing2/testuser@127.0.0.1:5433 closing because: unclean server (age=23)
	S-0x696e58: testing2/testuser@127.0.0.1:5433 new connection to server (from 127.0.0.1:38400)
	C-0x6921a8: testing2/testuser@1.2.3.4:33038 activate_client: finally got a connection

The new lines here are *per_loop_activate: reached per-user connection limit*
and *activate_client: finally got a connection*.

This introduces a fair bit of new code, including a new state in the state
machine (CL_WAITING_SLOT), so please review carefully.