Fix #41
As described in issue https://github.com/pgbouncer/pgbouncer/issues/41,
PgBouncer will fatally crash when config is reloaded that specifies a
user parameter for a database that already exists.

This is due to the config loader calling force_user for the newly
supplied user parameter, which creates a new user with an empty list of
user pools. When we later call find_pool(db, user) it is determined that
there is no associated pool, so we recreate a pool that already exists
and add it to the global pool list which then FATALs as a consequence of
duplicate elements.

We fix this bug by having force_user only create a new user if one for
this username does not already exist.