auth_dbname option for database definitions
Now you can specify an database's entry name for authentication.
If auth_dbname is set, auth_query is sended to that database.
This is util were we want use only one database for all users
authentications, i.e

   p0 = host=127.0.0.1 dbname=pgbench auth_dbname=authdb auth_user=pswcheck
   authdb = host=127.0.0.1 dbname=pgbouncer

If there isn't an entry named as the value specified in auth_dbname,
the database specified in dbname will be used instead.