Limit pgbouncer statistics to what the wire protocol supports.
Pgbouncer internal statistics are uint8 (64 bit), but the pg protocol is int8;
pgbouncer doesn't roll over the value it writes, instead assuming that the
wire protocol is uint when in fact it's int: see https://www.postgresql.org/docs/9.3/datatype-numeric.html .

To address this, the simplest fix is to force a modulus on what we serialize for uint8
values to limit them to int8 space.

It may be worth refactoring the internals of stats.c/admin.c rather than the
fix I'm proposing here- what is implemented here should cover all stats, but
it feels fragile in the sense that it's a duct tape fix tucked into an internal-
any consuming code that directly uses the underlying pktbuf functions can
bypass and reintroduce a variant of this bug.

I don't see a clean way to address this beyond vigilance combined w/ this
patch, thus proposing this fix.

Backing ticket for this is #150.