Extend max password length from 128 to 2048
When using AWS postgres RDS or Aurora it is possible to authenticate using a cryptographically signed token using IAM credentials. The password generated for connection is between typically about 800 characters long but can be in excess of 1300. The current buffer of 128 is inadequate and results in `WARNING bug in src/objects.c:461 - string truncated`
https://github.com/pgbouncer/pgbouncer/blob/dc226613ca023246684efc24b755e00b61ca6d84/src/objects.c#L449-L464