Fix issue with PAM users losing their password
The normal flow when starting a client connection is that set_pool()
is called with an empty password ("").  For normal users, the password
was already set earlier from auth_file or auth_query, so this empty
password is ignored.  For PAM users, the empty password is stored but
the real password is set later when authentication is complete.

The problem in the PAM case is that for the next client connection,
this would overwrite the stored password with an empty password until
the real password would then be re-added later.  If the client
authentication doesn't complete for whatever reason (perhaps server is
down and fast-fail is active), then the correct password is never set.
This would then have clobbered the user's password that might be
useful for server authentication.  There are probably other failure
scenarios.

To fix, call set_pool() with a NULL password instead, and teach
add_pam_user() not to overwrite an existing password if the argument
is NULL.

reported, analysis, and fix proposal by @achix

fixes #285