Fix some compiler warnings on Windows 64-bit
For event_set() callbacks, use evutil_socket_t instead of int as the
first argument.  This is required for 64-bit Windows; on all other
platforms, the two types are the same.

This requires libevent version 2.

Note that the rest of the socket code still uses int everywhere, so
Windows 64-bit is still probably not going to work correctly.