Fix reporting of adns implementation in configure
This was slightly broken by 748c83e5d8b05448725ce8a69213ed620a53d431
in the sense that configure would report both, say, c-ares and evdns
as "yes" without indication which one would actually be used.  Clean
that logic up a bit and report the actually used implementation at the
end of configure in a clearer way.  Also improve the documentation in
README.md around this.

Also clean up the build matrix in .travis.yml and remove a redundant
configuration.