Add so_reuseport option
Adds an option to specify whether to set the socket option
`SO_REUSEPORT` on TCP listening sockets.  On some operating systems,
this allows running multiple PgBouncer instances on the same host
listening on the same port and having the kernel distribute the
connections automatically.  This option is a way to get PgBouncer to
use more CPU cores.

closes #345