Fix issue #90: pgbouncer crashes if auto-added user is added to auth_file
When a user is auto-added to pgBouncer using `auth_query` (by `add_db_user`),
their pool is added to the global pool_list.  If the user is then added to
`auth_file` (e.g. because of #59), pgBouncer creates a new user object with the
same name (in `add_user`). Then, when it tries to find a pool for this user,
and doesn't (because they are newly created, so their own pool list is empty),
it calls new_pool to create one, which tries to add it to the global pool list
again (using `put_in_order`), which aborts (kills the process) because the same
username (different PgUser object) and DB name are already in the pool.

This patch appears to fix the crash, but is only lightly tested, and I'm not
sure whether we are supposed to remove the original user from the db list,
and/or update the global pool entry to point to the new PgUser object, or
whether it's OK to return a db pool entry from get_pool.