Fix test
The test_pool_size never actually worked correctly.  The way it was
written, the output of docount would collect the output of the select
pg_sleep calls, which would ultimately lead to an error from "test"
about too many arguments, but the way the test was written that would
result in a pass.

Fix that by rearranging the logic so that a syntax error from "test"
is a failure, and also hide the output of select pg_sleep so that the
test passes correctly.

Also run the pg_stat_activity probe query directly to the backend not
via PgBouncer so that it doesn't affect the pool sizes.  That doesn't
affect the validity of the test, but it seems cleaner.