Reorganize test log output
Previously, we had a per-test log file for the client
output (test_*.log) but a common pgbouncer log file for all
tests (test.log).  This was actually buggy: Some tests grepped the log
file for a certain error string, but nothing was ensuring that they
actually only checked the result of their own test, not a previous
test.

To fix this, move the pgbouncer log file out of the way after every
test and start a new one.  The saved per-test pgbouncer log files are
now test_*.log, the client output is put into test_*.out instead.