Test adjustment
Connect to 127.0.0.1 instead of localhost.  This avoids confusing
behavior when simulating connection failures.  When connecting to
localhost, psql/libpq would then, depending on circumstances, connect
to ::1 before reporting failure.  This also makes it more consistent
with the rest of the test setup, which uses mostly IP addresses.

For the ssl tests, we leave it as it was for now, since we don't have
any test cases there that are affected, and SSL is mostly host name
based in general.