Allow tests to run on platforms without getpeereid()
The tests require that psql can connect to the PgBouncer admin
console.  On platforms that have getpeereid(), this works by
connecting as user pgbouncer over the Unix socket.  On other
platforms, we have to rely on "trust" authentication, but then we have
to skip any tests that use authentication methods other than "trust".

This affects mostly Windows for now, but could possibly be applied to
other platforms as needed.