Allow running some of the test suite on Windows
With some hackery, we can run the main test script under
Windows/MinGW.  We don't use Unix-domain sockets (for now), we have to
skip tests that do authentication-related things, we have to start and
stop the daemon differently, and we can't use kill in shell scripts.
But the rest works, so at least we can test a fair amount of
functionality.

The other components of "make check" are not supported on Windows yet,
so for now we'll just run the test script directly on AppVeyor.