Mock authentication
On client login, when a user doesn't exist, don't respond with "no
such user", because we don't want hostile clients to be able to probe
what users exist.  Instead, play out the authentication exchange to
the end and then just respond that the authentication failed.  For
SCRAM, this requires a bit of extra work for constructing a fake
stored secret.  This all corresponds to how PostgreSQL works
internally as well.

This currently only works for users defined (or not) in auth_file.
For auth_query/auth_user, it doesn't work yet, for implementation
reasons that are described in code comments.  A workaround is possible
in some by defining auth_query in a way that it doesn't return no rows
when the user doesn't exist but instead returns a fake user record
with a wrong password, so that whatever password the client provided
would surely be rejected.