doc: Document that *_tls_ciphers does not affect TLS 1.3
TLS 1.3 uses a different way of specifying ciphers and a different
OpenSSL API.  PgBouncer currently does not support setting those
ciphers.  For now, just document this.  In the future, support for
this might be added somehow.  A similar resolution was also made in
PostgreSQL recently.

While we're here, also document the shortcut values available for
these settings.

closes #518