doc: Improve documentation of various logging parameters
Expand the documentation of how the various logging parameters and
options interact, such as that setting logfile does not turn off
logging to stderr and that setting logfile or syslog is required for
daemonization.

Also, the documentation was previously talking about stdout, but all
(non-file, non-syslog) logging actually goes to stderr, so change that
everywhere.

closes #492