Add warning message about systemd misconfiguration
Some distributions appear to have packaged pgbouncer in a way that the
systemd unit configuration expects a startup notification but
pgbouncer itself is built without systemd support.  To be able to
detect that situation better in the future, log a warning message at
startup if the environment variable NOTIFY_SOCKET is set but systemd
support was not built in.  Also add information about systemd support
to the --version output.