Fix behavior of global auth_user
A global auth_user setting was pretty confusing and fragile.  It was
looked up while the configuration was read, so it only worked if it
was set before the [databases] section.  Changing it at run time or
changing and reloading also worked in strange ways, depending on
circumstances, because of the ordering dependency.

To fix, look up the setting not while the configuration file is parsed
but later when we actually need it when authenticating a client
connection.

analysis and fix by @pinaraf (#393), some tweaking by me

closes #391