introduce a generation to a server connection
The generation is incremented for every command send to the server.
For cancelation requests we track the generation of a server connection.
By the time the requests reaches the front of the queue and a new connection is established the generation of a server is verified agains the generation of a cancelation request. If the generations don’t match the cancelation is aborted.

This mechanism prevents delayed cancelations from canceling the wrong statement executed after a connection has already sent new statements to the server.