Fix crash during mock authentication
When mock authentication is happening, the pool is not set, so there will
be a segmentation fault in check_fast_fail() when it is dereferenced.
This was not caught in the tests because they happened to use a forced
user, which ends up in a slightly different code path.  To improve
that, double the tests and make one with forced user and one without
for each covered authenticaton method.

fixes #541