Support hot reloading of TLS settings
This allows rotating TLS certificates without downtime. The reason this
change is so small is because OpenSSL already does reference counting of
the SSL context. All that's needed is load the values from the config
again and build a new SSL context that can be used for newly opened
connections. Any existing connections will continue using the old
context, which will get freed automatically when the reference count
hits zero.

closes #550, closes #403