Authentication for admin database
As the admin database is a "fake" one, auth_user is not being set for
it. Which in turn means if provided stats_user credentials should be put
into the auth_file, this user connected to the admin db can not be
authenticated via the underlying db itself.

In this case allow pgbouncer to connect to the default_auth_db to
perform authentication. This will allow to avoid managing another set of
credentials.