Set minimum pool size per database
This feature allows to set min_pool_size per database. The current
behavior uses the same min_pool_size setting for all databases. It could
not be suitable for some databases that has a different access pattern.

Let's say one database need a minimum number of connections to handle a
burst of connections in order to not increase the response time.
However, the other databases doesn't need such requirement. Hence, this
feature reduces the number of connections open by pgbouncer.

Closes #91