Make "unclean server" message clearer
The message "unclean server" is logged if the client connection is
closed while the server is not in "ready" state.  That can be if the
server is in a transaction block, because then we can't reset the
server to non-transaction state other than by closing it.  So this
message can happen in relatively normal use, and it's pretty unclear
and a bit scary.  So this change attempts to make the message a bit
more precise and document, at least in the code, what it means.