Allow cancel requests to bypass pool size
Before, when a pool is full, not even cancel requests can get through. This can be problematic in practice.

On the PostgreSQL side, cancel requests don't count against the max_connections limit. So it makes sense to make PgBouncer a bit more lenient as well. After some discussion, we didn't want to allow unlimited connections to be made, and we wanted to avoid adding another configuration setting. So as a heuristic, we allow cancel request connections up to twice the pool size.