Fix md5 authentication overwriting user name in memory
When using md5 authentication with a plain-text password, the old code
did the inner md5() call and overwrote the password in the in-memory
structures with that result.  That makes sense if md5 is the only
authentication method you ever use in an instance, because then you
don't need to do that call on every client login.  But with hba, a
different authentication method could be used next time, but this
behavior destroys the plain-text password.  For example, logging in
with md5 first will make a subsequent scram login fail, but not vice
versa.  To fix, don't cache the md5() result but just compute it every
time it is needed.  (We could in theory create a separate field for
caching the md5() output, but since this method is obsolescent, let's
not bother.)