Set LC_ALL in test scripts
The test script already set LANG, but that doesn't have any effect if
the user's environment sets an LC_* variable.  Because some of the
tests parse error messages (e.g., test_cancel*), we need to make sure
the user's locale doesn't interfere.

Also add this to the ssl test script.  At least that way the printed
timestamps come out in the same format.