Fixed multiple sequential queries logging
Running multiple queries in a pipelined way would trigger the dreaded
"FIXME: query end, but query_start == 0" log messages because when the
first query finishes, we reset the query_start, and then the return of
the second query gets upset.  To fix that, only compute the time when
all queries in flight have finished.

solution by @vonzshik

closes #39, closes #144, closes #215, closes #277, closes #376, closes #565