Send FATAL errors when client is to be disconnected
Although it is not technically part of the protocol specification,
some clients expect the connection to be closed after receiving a
FATAL error, and don't expect it to be closed after an ERROR-level
error.  So to be nice, use severity FATAL when we plan to close the
connection after sending the error.

based on patch by @x4m

closes #564