feat: introduce support for the `auth_dbname` option
Introduce a global or database-specific option (specified in the
connection parameters string) called `auth_dbname` which points
to the name of a database defined in the `[databases]` section
to be reserved for authentication purposes.

If required, the `auth_query` will be run against this database.

By default it is not set. It requires that `auth_user` is set.

Based on an early work by https://github.com/dbonne

Signed-off-by: Leonardo Cecchi <leonardo.cecchi@enterprisedb.com>
Signed-off-by: Gabriele Bartolini <gabriele.bartolini@enterprisedb.com>