Reject extraneous data after SSL or GSS encryption handshakes
PgBouncer, talking to a client (so PgBouncer is the server), collects
up to a bufferload of data whenever it reads data from the client
socket.  When SSL or GSS encryption is requested during startup, any
additional data received with the initial request message remained in
the buffer, and would be treated as already-decrypted data once the
encryption handshake completed.  Thus, a man-in-the-middle with the
ability to inject data into the TCP connection could stuff some
cleartext data into the start of a supposedly encryption-protected
database session.

This could be abused to send faked SQL commands to PgBouncer, although
it would only work if PgBouncer did not demand any authentication
data.  (However, an instance relying on SSL certificate authentication
might well not do so.)

To fix, throw a protocol-violation error if the internal buffer is not
empty after the encryption handshake.

(PgBouncer does not support GSS encryption, but we put the check in
place anyway for completeness.)

The analogous issue with PgBouncer talking to a server (so PgBouncer
is the client) did not exist, since there was already a check whether
the buffer was empty after reading an SSL response.  But just for
clarity, split out that check and give it its own error message.

Based on security report for PostgreSQL, reported by Jacob Champion,
code by Tom Lane; PgBouncer code by Peter Eisentraut, reviewed by
Andrew Dunstan