Clean up unportable sed use
Some test code was using sed -i in a GNU-specific way.  This led to
warnings when run with BSD sed.  There was probably no actual
misbehavior because those calls mainly addressed specific situations
that are expected to happen on systems with GNU sed.  In any case, it
turns out all these cases can be rewritten much simpler.  The main
change is that we no longer support running the test suite with
PostgreSQL versions with the old spelling "unix_socket_directory"
(singular), but we haven't actually supported such old releases in the
test suite for other reasons, so it's no loss.