Enable hba_test on Windows
The problem was that hba_test didn't build because it couldn't find
win32support.h.  To fix, rearrange how that header file is used.  It
is really specific to pgbouncer's main invocation, so it shouldn't be
included by other programs.  So remove it from system.h and include it
directly from main.c.  This fixes the build of hba_test.  Now we can
invoke "make check" on Windows (but without the SSL tests for now,
which need more work).