Mock authentication for missing databases
On client login, when a database does not exist, don't immediately
respond with "no such database", because we don't want hostile clients
to be able to probe what databases exist.  Instead, do the
authentication first and then respond about the database.

This is similar to 7f7ff12a4288d9ab52154c7229afc0186f8196bb, but for
databases.