Call cleanup() via atexit()
We cannot call cleanup() from main() because there are atexit()
handlers such as cleanup_sockets() that still need the configuration
settings.  This currently leads to a crash if you stop pgbouncer using
SIGINT when assertions are enabled.  To fix, make cleanup() an
atexit() handler also and make sure it is called last.

closes #685