Improve tcp_defer_accept documentation and behavior
The documentation was incorrect and misleading about the default.
Also, in some cases the wrong value was showing in "show config".

While we are here, improve the implementation a bit: If it's set but
not supported, give an error, similar to how other platform-specific
socket options are handled.  Finally, remove the long-disabled
SO_ACCEPTFILTER code.

closes #686