Fix c-ares header file handling
As of version 1.18.0, c-ares provides ares_nameser.h, which is a
portability substitute/wrapper for arpa/nameser.h.  This is needed
mainly on Windows, where the latter does not exist.  So we add a
configure check for ares_nameser.h and include it if it exists instead
of arpa/nameser.h.

This alters the change made by
ad27b82d09f4332c707580a37bc2f9d6fa3f867e: There it was claimed that
c-ares provides nameser.h, but that header file was only internal and
never actually officially installed by c-ares.  Since there is now an
official solution for this, let's not keep maintaining support for
some short-lived unofficial solution.

The upshot is: On Windows, PgBouncer (now) requires c-ares >=1.18.0.

closes #657