Cirrus CI: Organize Ubuntu/Debian tasks better
Reduce the number of tasks by combining some orthogonal configure
options.  Also, run all Ubuntu/Debian tasks with systemd support now.
It has turned out to be important in practice to check with a variety
of old and new systemd versions.  The non-systemd code paths are
already tested on non-Linux builds (currently Windows).

Set the standard PostgreSQL version to 14, and have the various tasks
test various versions so that we get some reasonable coverage of
different versions.  (The version chosen in each case is mostly close
to what the default version was for that distribution.)

At the moment, PostgreSQL versions relevant for PgBouncer testing are:
9.6 because it is the most recent version without SCRAM support, and
13 because it is the most recent version with a limited maximum
password length.