Remove automatic auth_file reload
Before, the auth_file was stat()ed three times a second and
automatically reloaded when it was changed.  This caused some
performance problems if the file system is very busy, and it was also
undocumented and surprising.

Now, the auth_file is only reloaded when a configuration file reload
is explicitly requested (RELOAD command or similar).  This is also
documented now.

Presumably, the origin of this behavior is that in ancient times,
auth_file was meant to point into the PostgreSQL data directory, and
with this behavior it would pick up user changes made through
PostgreSQL automatically.  But that is no longer the way it is used,
so an explicit reload is more appropriate now.

closes #377, closes #574