Don't send server disconnect errors to client before login
disconnect_server() also sends the error message to the client if the
server is linked.  However, this can reveal not-quite-public
information, such as configuration details, to a client that is not
logged in yet.  So change this to send the message only if the client
has completed login, and send a generic message otherwise.