Docs: Clarify how summary client wait times are calculated
As is these docs can be quite confusing since total_wait_time doesn't
increase in parallel with obviously long-waiting clients. Worse,
avg_wait_time is effectively not a meaningful metric given that it
calculates wait time averaged across the seconds in a stat period and
also doesn't include anything until a client is assigned a backend.
Probably that should be fixed (and/or the current implementation
deprecated), but until then we should at least document it.