Clean up waiting cancel request in janitor (#786)
While doing some final testing of cancellations after merging #717 I
found a small issue. Query cancellations that are waiting to receive a
server connection never timeout. This behavior wasn't great before #717,
but after #717 it became a bit worse. Because pgbouncer now keeps the
client connection open until the cancellation is actually acknowledged
by postgres.

Because psql sends cancellations in a blocking way in a signal handler,
this can actually create a type of deadlock when sending SIGINT to
pgbouncer while psql is trying to cancel a query. When sending SIGINT to
pgbouncer, it will wait for all query clients to be finished before
turning off. But the cancel request will be indefinitely wait for a new
connection. And while it's doing that psql will never consume the output
from the already runinng query, and thus pgbouncer will stay up until
the SHUTDOWN command is sent to resolve this deadlock.

Normal connections that are waiting for a server connection will timeout
based on query_timeout and query_wait_timeout. This starts using these
same timeouts for waiting cancel connections. This resolves this issue
by breaking the infinite wait in libpq.