Rename wait_cancels state to being_canceled (#788)
After merging #717 it was noted that the `wait_cancels` name was quite
confusing. This changes that name to `being_canceled`. This makes it
more clear that this is not a server that is canceling itself, but
instead it's the target of a cancellation request.


Another option would be to rename it to waiting_cancel to be symetric 
with the client state. However, not being symetric is kind of the point 
of the newly suggested name. 

sv_wait_cancel would suggest that it's waiting for a cancel that it sent itself:
- active_cancel_req_list, is a list of clients that are sending a cancel themselves
- waiting_cancel_req_list, is a list of clients that are waiting for a server 
  connection needed to for the cancel
- active_cancel_server_list, is the list of server connections on which a cancel
  is currently being sent

Thus (to me) waiting_cancel_server_list would suggest that any server 
in this list is waiting to send a cancel. But that is not what it means 
to be in this list. Servers in the list in question are currently being 
cancelled by another server connection. To make this difference
clear is why I suggested `being_cancelled_server_list`.